move common functions to displayUtils
